updated migrate

+++ database/migrations/2025_08_13_103133_remove_short_description_from_services_table.php
@@ -12,12 +12,13 @@ return new class extends Migration
      * @return void
      */
     public function up()
-    {
+{
+    if (Schema::hasColumn('services', 'short_description')) {
         Schema::table('services', function (Blueprint $table) {
             $table->dropColumn('short_description');
-
         });
     }
+}
 
     /**
      * Reverse the migrations.
